+++ mobile/android/android-components/docs/changelog.md
@@ -11,6 +11,10 @@ permalink: /changelog/
 * [Gecko](https://github.com/mozilla-mobile/android-components/blob/main/buildSrc/src/main/java/Gecko.kt)
 * [Configuration](https://github.com/mozilla-mobile/android-components/blob/main/.config.yml)
 
+* **service-firefox-accounts**
+  * `SyncStatus` can now be `LoggedOut`.
+  * `SyncStoreSupport` will update the `SyncStore` with `LoggedOut` when observed.
+
 * **feature-recentlyclosed**
   * 🚒 Bug fixed [issue #12470](https://github.com/mozilla-mobile/android-components/issues/12470) - Set autoMirrored to true to fix RTL issues